bot/requirements.txt
python-dotenv
python-telegram-bot==20.4
requests
PyJWT